LayoutTests/inspector/formatting/source-map-javascript-3.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/source-map-utilities.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("SourceMaps.JavaScript");
    addSourceMapTest(suite, "text/javascript", "3", "resources/javascript-source-map-tests/3.js");
    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test JavaScript SourceMap with multi-line JS template literals</p>
</body>
</html>
