LayoutTests/inspector/formatting/source-map-javascript-3-expected.txt

Test JavaScript SourceMap with multi-line JS template literals


== Running test suite: SourceMaps.JavaScript
-- Running test case: SourceMaps.JavaScript.3
Original Content Length
Formatted Content Length
---- Original to Formatted ----
Original: 0 (0, 1)
  0| `#A
  1| B
  2| C`;
Formatted: 0 (0, 1)
  0| `#A
  1| B
  2| C`;

Original: 1 (0, 2)
  0| `A#
  1| B
  2| C`;
Formatted: 1 (0, 2)
  0| `A#
  1| B
  2| C`;

Original: 2 (1, 0)
  0| `A
  1| #B
  2| C`;
  3| let a=1,b=2,c=3;
Formatted: 2 (1, 0)
  0| `A
  1| #B
  2| C`;
  3| let a = 1,

Original: 3 (1, 1)
  0| `A
  1| B#
  2| C`;
  3| let a=1,b=2,c=3;
Formatted: 3 (1, 1)
  0| `A
  1| B#
  2| C`;
  3| let a = 1,

Original: 4 (2, 0)
  0| `A
  1| B
  2| #C`;
  3| let a=1,b=2,c=3;
  4| test`A
Formatted: 4 (2, 0)
  0| `A
  1| B
  2| #C`;
  3| let a = 1,
  4|     b = 2,

Original: 5 (2, 1)
  0| `A
  1| B
  2| C#`;
  3| let a=1,b=2,c=3;
  4| test`A
Formatted: 5 (2, 1)
  0| `A
  1| B
  2| C#`;
  3| let a = 1,
  4|     b = 2,

Original: 6 (2, 2)
  0| `A
  1| B
  2| C`#;
  3| let a=1,b=2,c=3;
  4| test`A
Formatted: 6 (2, 2)
  0| `A
  1| B
  2| C`#;
  3| let a = 1,
  4|     b = 2,

Original: 7 (2, 3)
  0| `A
  1| B
  2| C`;#
  3| let a=1,b=2,c=3;
  4| test`A
Formatted: 7 (2, 3)
  0| `A
  1| B
  2| C`;#
  3| let a = 1,
  4|     b = 2,

Original: 8 (3, 0)
  1| B
  2| C`;
  3| #let a=1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 8 (3, 0)
  1| B
  2| C`;
  3| #let a = 1,
  4|     b = 2,
  5|     c = 3;

Original: 9 (3, 1)
  1| B
  2| C`;
  3| l#et a=1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 9 (3, 1)
  1| B
  2| C`;
  3| l#et a = 1,
  4|     b = 2,
  5|     c = 3;

Original: 10 (3, 2)
  1| B
  2| C`;
  3| le#t a=1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 10 (3, 2)
  1| B
  2| C`;
  3| le#t a = 1,
  4|     b = 2,
  5|     c = 3;

Original: 11 (3, 3)
  1| B
  2| C`;
  3| let# a=1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 11 (3, 3)
  1| B
  2| C`;
  3| let# a = 1,
  4|     b = 2,
  5|     c = 3;

Original: 12 (3, 4)
  1| B
  2| C`;
  3| let #a=1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 12 (3, 4)
  1| B
  2| C`;
  3| let #a = 1,
  4|     b = 2,
  5|     c = 3;

Original: 13 (3, 5)
  1| B
  2| C`;
  3| let a#=1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 13 (3, 6)
  1| B
  2| C`;
  3| let a #= 1,
  4|     b = 2,
  5|     c = 3;

Original: 14 (3, 6)
  1| B
  2| C`;
  3| let a=#1,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 15 (3, 8)
  1| B
  2| C`;
  3| let a = #1,
  4|     b = 2,
  5|     c = 3;

Original: 15 (3, 7)
  1| B
  2| C`;
  3| let a=1#,b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 17 (3, 9)
  1| B
  2| C`;
  3| let a = 1#,
  4|     b = 2,
  5|     c = 3;

Original: 16 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}
Formatted: 18 (4, 4)
  2| C`;
  3| let a = 1,
  4|     #b = 2,
  5|     c = 3;
  6| test`A

Original: 17 (3, 9)
  1| B
  2| C`;
  3| let a=1,b#=2,c=3;
  4| test`A
  5| ${a}
Formatted: 24 (4, 6)
  2| C`;
  3| let a = 1,
  4|     b #= 2,
  5|     c = 3;
  6| test`A

Original: 18 (3, 10)
  1| B
  2| C`;
  3| let a=1,b=#2,c=3;
  4| test`A
  5| ${a}
Formatted: 26 (4, 8)
  2| C`;
  3| let a = 1,
  4|     b = #2,
  5|     c = 3;
  6| test`A

Original: 19 (3, 11)
  1| B
  2| C`;
  3| let a=1,b=2#,c=3;
  4| test`A
  5| ${a}
Formatted: 28 (4, 9)
  2| C`;
  3| let a = 1,
  4|     b = 2#,
  5|     c = 3;
  6| test`A

Original: 20 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}
Formatted: 29 (5, 4)
  3| let a = 1,
  4|     b = 2,
  5|     #c = 3;
  6| test`A
  7| ${a}

Original: 21 (3, 13)
  1| B
  2| C`;
  3| let a=1,b=2,c#=3;
  4| test`A
  5| ${a}
Formatted: 35 (5, 6)
  3| let a = 1,
  4|     b = 2,
  5|     c #= 3;
  6| test`A
  7| ${a}

Original: 22 (3, 14)
  1| B
  2| C`;
  3| let a=1,b=2,c=#3;
  4| test`A
  5| ${a}
Formatted: 37 (5, 8)
  3| let a = 1,
  4|     b = 2,
  5|     c = #3;
  6| test`A
  7| ${a}

Original: 23 (3, 15)
  1| B
  2| C`;
  3| let a=1,b=2,c=3#;
  4| test`A
  5| ${a}
Formatted: 39 (5, 9)
  3| let a = 1,
  4|     b = 2,
  5|     c = 3#;
  6| test`A
  7| ${a}

Original: 24 (3, 16)
  1| B
  2| C`;
  3| let a=1,b=2,c=3;#
  4| test`A
  5| ${a}
Formatted: 40 (5, 10)
  3| let a = 1,
  4|     b = 2,
  5|     c = 3;#
  6| test`A
  7| ${a}

Original: 25 (4, 0)
  2| C`;
  3| let a=1,b=2,c=3;
  4| #test`A
  5| ${a}
  6| B
Formatted: 41 (6, 0)
  4|     b = 2,
  5|     c = 3;
  6| #test`A
  7| ${a}
  8| B

Original: 26 (4, 1)
  2| C`;
  3| let a=1,b=2,c=3;
  4| t#est`A
  5| ${a}
  6| B
Formatted: 42 (6, 1)
  4|     b = 2,
  5|     c = 3;
  6| t#est`A
  7| ${a}
  8| B

Original: 27 (4, 2)
  2| C`;
  3| let a=1,b=2,c=3;
  4| te#st`A
  5| ${a}
  6| B
Formatted: 43 (6, 2)
  4|     b = 2,
  5|     c = 3;
  6| te#st`A
  7| ${a}
  8| B

Original: 28 (4, 3)
  2| C`;
  3| let a=1,b=2,c=3;
  4| tes#t`A
  5| ${a}
  6| B
Formatted: 44 (6, 3)
  4|     b = 2,
  5|     c = 3;
  6| tes#t`A
  7| ${a}
  8| B

Original: 29 (4, 4)
  2| C`;
  3| let a=1,b=2,c=3;
  4| test#`A
  5| ${a}
  6| B
Formatted: 45 (6, 4)
  4|     b = 2,
  5|     c = 3;
  6| test#`A
  7| ${a}
  8| B

Original: 30 (4, 5)
  2| C`;
  3| let a=1,b=2,c=3;
  4| test`#A
  5| ${a}
  6| B
Formatted: 46 (6, 5)
  4|     b = 2,
  5|     c = 3;
  6| test`#A
  7| ${a}
  8| B

Original: 31 (4, 6)
  2| C`;
  3| let a=1,b=2,c=3;
  4| test`A#
  5| ${a}
  6| B
Formatted: 47 (6, 6)
  4|     b = 2,
  5|     c = 3;
  6| test`A#
  7| ${a}
  8| B

Original: 32 (5, 0)
  3| let a=1,b=2,c=3;
  4| test`A
  5| #${a}
  6| B
  7| ${b}
Formatted: 48 (7, 0)
  5|     c = 3;
  6| test`A
  7| #${a}
  8| B
  9| ${b}

Original: 33 (5, 1)
  3| let a=1,b=2,c=3;
  4| test`A
  5| $#{a}
  6| B
  7| ${b}
Formatted: 49 (7, 1)
  5|     c = 3;
  6| test`A
  7| $#{a}
  8| B
  9| ${b}

Original: 34 (5, 2)
  3| let a=1,b=2,c=3;
  4| test`A
  5| ${#a}
  6| B
  7| ${b}
Formatted: 50 (7, 2)
  5|     c = 3;
  6| test`A
  7| ${#a}
  8| B
  9| ${b}

Original: 35 (5, 3)
  3| let a=1,b=2,c=3;
  4| test`A
  5| ${a#}
  6| B
  7| ${b}
Formatted: 51 (7, 3)
  5|     c = 3;
  6| test`A
  7| ${a#}
  8| B
  9| ${b}

Original: 36 (5, 4)
  3| let a=1,b=2,c=3;
  4| test`A
  5| ${a}#
  6| B
  7| ${b}
Formatted: 52 (7, 4)
  5|     c = 3;
  6| test`A
  7| ${a}#
  8| B
  9| ${b}

Original: 37 (6, 0)
  4| test`A
  5| ${a}
  6| #B
  7| ${b}
  8| C
Formatted: 53 (8, 0)
  6| test`A
  7| ${a}
  8| #B
  9| ${b}
 10| C

Original: 38 (6, 1)
  4| test`A
  5| ${a}
  6| B#
  7| ${b}
  8| C
Formatted: 54 (8, 1)
  6| test`A
  7| ${a}
  8| B#
  9| ${b}
 10| C

Original: 39 (7, 0)
  5| ${a}
  6| B
  7| #${b}
  8| C
  9| ${c}`;
Formatted: 55 (9, 0)
  7| ${a}
  8| B
  9| #${b}
 10| C
 11| ${c}`;

Original: 40 (7, 1)
  5| ${a}
  6| B
  7| $#{b}
  8| C
  9| ${c}`;
Formatted: 56 (9, 1)
  7| ${a}
  8| B
  9| $#{b}
 10| C
 11| ${c}`;

Original: 41 (7, 2)
  5| ${a}
  6| B
  7| ${#b}
  8| C
  9| ${c}`;
Formatted: 57 (9, 2)
  7| ${a}
  8| B
  9| ${#b}
 10| C
 11| ${c}`;

Original: 42 (7, 3)
  5| ${a}
  6| B
  7| ${b#}
  8| C
  9| ${c}`;
Formatted: 58 (9, 3)
  7| ${a}
  8| B
  9| ${b#}
 10| C
 11| ${c}`;

Original: 43 (7, 4)
  5| ${a}
  6| B
  7| ${b}#
  8| C
  9| ${c}`;
Formatted: 59 (9, 4)
  7| ${a}
  8| B
  9| ${b}#
 10| C
 11| ${c}`;

Original: 44 (8, 0)
  6| B
  7| ${b}
  8| #C
  9| ${c}`;
 10| `ABC`;
Formatted: 60 (10, 0)
  8| B
  9| ${b}
 10| #C
 11| ${c}`;
 12| `ABC`;

Original: 45 (8, 1)
  6| B
  7| ${b}
  8| C#
  9| ${c}`;
 10| `ABC`;
Formatted: 61 (10, 1)
  8| B
  9| ${b}
 10| C#
 11| ${c}`;
 12| `ABC`;

Original: 46 (9, 0)
  7| ${b}
  8| C
  9| #${c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 62 (11, 0)
  9| ${b}
 10| C
 11| #${c}`;
 12| `ABC`;
 13| a = 1,

Original: 47 (9, 1)
  7| ${b}
  8| C
  9| $#{c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 63 (11, 1)
  9| ${b}
 10| C
 11| $#{c}`;
 12| `ABC`;
 13| a = 1,

Original: 48 (9, 2)
  7| ${b}
  8| C
  9| ${#c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 64 (11, 2)
  9| ${b}
 10| C
 11| ${#c}`;
 12| `ABC`;
 13| a = 1,

Original: 49 (9, 3)
  7| ${b}
  8| C
  9| ${c#}`;
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 65 (11, 3)
  9| ${b}
 10| C
 11| ${c#}`;
 12| `ABC`;
 13| a = 1,

Original: 50 (9, 4)
  7| ${b}
  8| C
  9| ${c}#`;
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 66 (11, 4)
  9| ${b}
 10| C
 11| ${c}#`;
 12| `ABC`;
 13| a = 1,

Original: 51 (9, 5)
  7| ${b}
  8| C
  9| ${c}`#;
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 67 (11, 5)
  9| ${b}
 10| C
 11| ${c}`#;
 12| `ABC`;
 13| a = 1,

Original: 52 (9, 6)
  7| ${b}
  8| C
  9| ${c}`;#
 10| `ABC`;
 11| a=1,b=2,c=3;
Formatted: 68 (11, 6)
  9| ${b}
 10| C
 11| ${c}`;#
 12| `ABC`;
 13| a = 1,

Original: 53 (10, 0)
  8| C
  9| ${c}`;
 10| #`ABC`;
 11| a=1,b=2,c=3;
 12|
Formatted: 69 (12, 0)
 10| C
 11| ${c}`;
 12| #`ABC`;
 13| a = 1,
 14| b = 2,

Original: 54 (10, 1)
  8| C
  9| ${c}`;
 10| `#ABC`;
 11| a=1,b=2,c=3;
 12|
Formatted: 70 (12, 1)
 10| C
 11| ${c}`;
 12| `#ABC`;
 13| a = 1,
 14| b = 2,

Original: 55 (10, 2)
  8| C
  9| ${c}`;
 10| `A#BC`;
 11| a=1,b=2,c=3;
 12|
Formatted: 71 (12, 2)
 10| C
 11| ${c}`;
 12| `A#BC`;
 13| a = 1,
 14| b = 2,

Original: 56 (10, 3)
  8| C
  9| ${c}`;
 10| `AB#C`;
 11| a=1,b=2,c=3;
 12|
Formatted: 72 (12, 3)
 10| C
 11| ${c}`;
 12| `AB#C`;
 13| a = 1,
 14| b = 2,

Original: 57 (10, 4)
  8| C
  9| ${c}`;
 10| `ABC#`;
 11| a=1,b=2,c=3;
 12|
Formatted: 73 (12, 4)
 10| C
 11| ${c}`;
 12| `ABC#`;
 13| a = 1,
 14| b = 2,

Original: 58 (10, 5)
  8| C
  9| ${c}`;
 10| `ABC`#;
 11| a=1,b=2,c=3;
 12|
Formatted: 74 (12, 5)
 10| C
 11| ${c}`;
 12| `ABC`#;
 13| a = 1,
 14| b = 2,

Original: 59 (10, 6)
  8| C
  9| ${c}`;
 10| `ABC`;#
 11| a=1,b=2,c=3;
 12|
Formatted: 75 (12, 6)
 10| C
 11| ${c}`;
 12| `ABC`;#
 13| a = 1,
 14| b = 2,

Original: 60 (11, 0)
  9| ${c}`;
 10| `ABC`;
 11| #a=1,b=2,c=3;
 12|
Formatted: 76 (13, 0)
 11| ${c}`;
 12| `ABC`;
 13| #a = 1,
 14| b = 2,
 15| c = 3;

Original: 61 (11, 1)
  9| ${c}`;
 10| `ABC`;
 11| a#=1,b=2,c=3;
 12|
Formatted: 77 (13, 2)
 11| ${c}`;
 12| `ABC`;
 13| a #= 1,
 14| b = 2,
 15| c = 3;

Original: 62 (11, 2)
  9| ${c}`;
 10| `ABC`;
 11| a=#1,b=2,c=3;
 12|
Formatted: 79 (13, 4)
 11| ${c}`;
 12| `ABC`;
 13| a = #1,
 14| b = 2,
 15| c = 3;

Original: 63 (11, 3)
  9| ${c}`;
 10| `ABC`;
 11| a=1#,b=2,c=3;
 12|
Formatted: 81 (13, 5)
 11| ${c}`;
 12| `ABC`;
 13| a = 1#,
 14| b = 2,
 15| c = 3;

Original: 64 (11, 4)
  9| ${c}`;
 10| `ABC`;
 11| a=1,#b=2,c=3;
 12|
Formatted: 82 (14, 0)
 12| `ABC`;
 13| a = 1,
 14| #b = 2,
 15| c = 3;
 16|

Original: 65 (11, 5)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b#=2,c=3;
 12|
Formatted: 84 (14, 2)
 12| `ABC`;
 13| a = 1,
 14| b #= 2,
 15| c = 3;
 16|

Original: 66 (11, 6)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=#2,c=3;
 12|
Formatted: 86 (14, 4)
 12| `ABC`;
 13| a = 1,
 14| b = #2,
 15| c = 3;
 16|

Original: 67 (11, 7)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2#,c=3;
 12|
Formatted: 88 (14, 5)
 12| `ABC`;
 13| a = 1,
 14| b = 2#,
 15| c = 3;
 16|

Original: 68 (11, 8)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,#c=3;
 12|
Formatted: 89 (15, 0)
 13| a = 1,
 14| b = 2,
 15| #c = 3;
 16|

Original: 69 (11, 9)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c#=3;
 12|
Formatted: 91 (15, 2)
 13| a = 1,
 14| b = 2,
 15| c #= 3;
 16|

Original: 70 (11, 10)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=#3;
 12|
Formatted: 93 (15, 4)
 13| a = 1,
 14| b = 2,
 15| c = #3;
 16|

Original: 71 (11, 11)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=3#;
 12|
Formatted: 95 (15, 5)
 13| a = 1,
 14| b = 2,
 15| c = 3#;
 16|

Original: 72 (11, 12)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;#
 12|
Formatted: 96 (15, 6)
 13| a = 1,
 14| b = 2,
 15| c = 3;#
 16|

Original: 73 (12, 0)
 10| `ABC`;
 11| a=1,b=2,c=3;
 12| #
Formatted: 97 (16, 0)
 14| b = 2,
 15| c = 3;
 16| #


---- Formatted to Original ----
Formatted: 0 (0, 1)
  0| `#A
  1| B
  2| C`;
Original: 0 (0, 1)
  0| `#A
  1| B
  2| C`;

Formatted: 1 (0, 2)
  0| `A#
  1| B
  2| C`;
Original: 1 (0, 2)
  0| `A#
  1| B
  2| C`;

Formatted: 2 (1, 0)
  0| `A
  1| #B
  2| C`;
  3| let a = 1,
Original: 2 (1, 0)
  0| `A
  1| #B
  2| C`;
  3| let a=1,b=2,c=3;

Formatted: 3 (1, 1)
  0| `A
  1| B#
  2| C`;
  3| let a = 1,
Original: 3 (1, 1)
  0| `A
  1| B#
  2| C`;
  3| let a=1,b=2,c=3;

Formatted: 4 (2, 0)
  0| `A
  1| B
  2| #C`;
  3| let a = 1,
  4|     b = 2,
Original: 4 (2, 0)
  0| `A
  1| B
  2| #C`;
  3| let a=1,b=2,c=3;
  4| test`A

Formatted: 5 (2, 1)
  0| `A
  1| B
  2| C#`;
  3| let a = 1,
  4|     b = 2,
Original: 5 (2, 1)
  0| `A
  1| B
  2| C#`;
  3| let a=1,b=2,c=3;
  4| test`A

Formatted: 6 (2, 2)
  0| `A
  1| B
  2| C`#;
  3| let a = 1,
  4|     b = 2,
Original: 6 (2, 2)
  0| `A
  1| B
  2| C`#;
  3| let a=1,b=2,c=3;
  4| test`A

Formatted: 7 (2, 3)
  0| `A
  1| B
  2| C`;#
  3| let a = 1,
  4|     b = 2,
Original: 7 (2, 3)
  0| `A
  1| B
  2| C`;#
  3| let a=1,b=2,c=3;
  4| test`A

Formatted: 8 (3, 0)
  1| B
  2| C`;
  3| #let a = 1,
  4|     b = 2,
  5|     c = 3;
Original: 8 (3, 0)
  1| B
  2| C`;
  3| #let a=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 9 (3, 1)
  1| B
  2| C`;
  3| l#et a = 1,
  4|     b = 2,
  5|     c = 3;
Original: 9 (3, 1)
  1| B
  2| C`;
  3| l#et a=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 10 (3, 2)
  1| B
  2| C`;
  3| le#t a = 1,
  4|     b = 2,
  5|     c = 3;
Original: 10 (3, 2)
  1| B
  2| C`;
  3| le#t a=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 11 (3, 3)
  1| B
  2| C`;
  3| let# a = 1,
  4|     b = 2,
  5|     c = 3;
Original: 11 (3, 3)
  1| B
  2| C`;
  3| let# a=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 12 (3, 4)
  1| B
  2| C`;
  3| let #a = 1,
  4|     b = 2,
  5|     c = 3;
Original: 12 (3, 4)
  1| B
  2| C`;
  3| let #a=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 13 (3, 5)
  1| B
  2| C`;
  3| let a# = 1,
  4|     b = 2,
  5|     c = 3;
Original: 13 (3, 5)
  1| B
  2| C`;
  3| let a#=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 14 (3, 6)
  1| B
  2| C`;
  3| let a #= 1,
  4|     b = 2,
  5|     c = 3;
Original: 14 (3, 5)
  1| B
  2| C`;
  3| let a#=1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 15 (3, 7)
  1| B
  2| C`;
  3| let a =# 1,
  4|     b = 2,
  5|     c = 3;
Original: 14 (3, 6)
  1| B
  2| C`;
  3| let a=#1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 16 (3, 8)
  1| B
  2| C`;
  3| let a = #1,
  4|     b = 2,
  5|     c = 3;
Original: 15 (3, 6)
  1| B
  2| C`;
  3| let a=#1,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 17 (3, 9)
  1| B
  2| C`;
  3| let a = 1#,
  4|     b = 2,
  5|     c = 3;
Original: 15 (3, 7)
  1| B
  2| C`;
  3| let a=1#,b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 18 (3, 10)
  1| B
  2| C`;
  3| let a = 1,#
  4|     b = 2,
  5|     c = 3;
Original: 16 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 19 (4, 0)
  2| C`;
  3| let a = 1,
  4| #    b = 2,
  5|     c = 3;
  6| test`A
Original: 17 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 20 (4, 1)
  2| C`;
  3| let a = 1,
  4|  #   b = 2,
  5|     c = 3;
  6| test`A
Original: 17 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 21 (4, 2)
  2| C`;
  3| let a = 1,
  4|   #  b = 2,
  5|     c = 3;
  6| test`A
Original: 17 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 22 (4, 3)
  2| C`;
  3| let a = 1,
  4|    # b = 2,
  5|     c = 3;
  6| test`A
Original: 17 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 23 (4, 4)
  2| C`;
  3| let a = 1,
  4|     #b = 2,
  5|     c = 3;
  6| test`A
Original: 17 (3, 8)
  1| B
  2| C`;
  3| let a=1,#b=2,c=3;
  4| test`A
  5| ${a}

Formatted: 24 (4, 5)
  2| C`;
  3| let a = 1,
  4|     b# = 2,
  5|     c = 3;
  6| test`A
Original: 17 (3, 9)
  1| B
  2| C`;
  3| let a=1,b#=2,c=3;
  4| test`A
  5| ${a}

Formatted: 25 (4, 6)
  2| C`;
  3| let a = 1,
  4|     b #= 2,
  5|     c = 3;
  6| test`A
Original: 18 (3, 9)
  1| B
  2| C`;
  3| let a=1,b#=2,c=3;
  4| test`A
  5| ${a}

Formatted: 26 (4, 7)
  2| C`;
  3| let a = 1,
  4|     b =# 2,
  5|     c = 3;
  6| test`A
Original: 18 (3, 10)
  1| B
  2| C`;
  3| let a=1,b=#2,c=3;
  4| test`A
  5| ${a}

Formatted: 27 (4, 8)
  2| C`;
  3| let a = 1,
  4|     b = #2,
  5|     c = 3;
  6| test`A
Original: 19 (3, 10)
  1| B
  2| C`;
  3| let a=1,b=#2,c=3;
  4| test`A
  5| ${a}

Formatted: 28 (4, 9)
  2| C`;
  3| let a = 1,
  4|     b = 2#,
  5|     c = 3;
  6| test`A
Original: 19 (3, 11)
  1| B
  2| C`;
  3| let a=1,b=2#,c=3;
  4| test`A
  5| ${a}

Formatted: 29 (4, 10)
  2| C`;
  3| let a = 1,
  4|     b = 2,#
  5|     c = 3;
  6| test`A
Original: 20 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}

Formatted: 30 (5, 0)
  3| let a = 1,
  4|     b = 2,
  5| #    c = 3;
  6| test`A
  7| ${a}
Original: 21 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}

Formatted: 31 (5, 1)
  3| let a = 1,
  4|     b = 2,
  5|  #   c = 3;
  6| test`A
  7| ${a}
Original: 21 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}

Formatted: 32 (5, 2)
  3| let a = 1,
  4|     b = 2,
  5|   #  c = 3;
  6| test`A
  7| ${a}
Original: 21 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}

Formatted: 33 (5, 3)
  3| let a = 1,
  4|     b = 2,
  5|    # c = 3;
  6| test`A
  7| ${a}
Original: 21 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}

Formatted: 34 (5, 4)
  3| let a = 1,
  4|     b = 2,
  5|     #c = 3;
  6| test`A
  7| ${a}
Original: 21 (3, 12)
  1| B
  2| C`;
  3| let a=1,b=2,#c=3;
  4| test`A
  5| ${a}

Formatted: 35 (5, 5)
  3| let a = 1,
  4|     b = 2,
  5|     c# = 3;
  6| test`A
  7| ${a}
Original: 21 (3, 13)
  1| B
  2| C`;
  3| let a=1,b=2,c#=3;
  4| test`A
  5| ${a}

Formatted: 36 (5, 6)
  3| let a = 1,
  4|     b = 2,
  5|     c #= 3;
  6| test`A
  7| ${a}
Original: 22 (3, 13)
  1| B
  2| C`;
  3| let a=1,b=2,c#=3;
  4| test`A
  5| ${a}

Formatted: 37 (5, 7)
  3| let a = 1,
  4|     b = 2,
  5|     c =# 3;
  6| test`A
  7| ${a}
Original: 22 (3, 14)
  1| B
  2| C`;
  3| let a=1,b=2,c=#3;
  4| test`A
  5| ${a}

Formatted: 38 (5, 8)
  3| let a = 1,
  4|     b = 2,
  5|     c = #3;
  6| test`A
  7| ${a}
Original: 23 (3, 14)
  1| B
  2| C`;
  3| let a=1,b=2,c=#3;
  4| test`A
  5| ${a}

Formatted: 39 (5, 9)
  3| let a = 1,
  4|     b = 2,
  5|     c = 3#;
  6| test`A
  7| ${a}
Original: 23 (3, 15)
  1| B
  2| C`;
  3| let a=1,b=2,c=3#;
  4| test`A
  5| ${a}

Formatted: 40 (5, 10)
  3| let a = 1,
  4|     b = 2,
  5|     c = 3;#
  6| test`A
  7| ${a}
Original: 24 (3, 16)
  1| B
  2| C`;
  3| let a=1,b=2,c=3;#
  4| test`A
  5| ${a}

Formatted: 41 (6, 0)
  4|     b = 2,
  5|     c = 3;
  6| #test`A
  7| ${a}
  8| B
Original: 25 (4, 0)
  2| C`;
  3| let a=1,b=2,c=3;
  4| #test`A
  5| ${a}
  6| B

Formatted: 42 (6, 1)
  4|     b = 2,
  5|     c = 3;
  6| t#est`A
  7| ${a}
  8| B
Original: 26 (4, 1)
  2| C`;
  3| let a=1,b=2,c=3;
  4| t#est`A
  5| ${a}
  6| B

Formatted: 43 (6, 2)
  4|     b = 2,
  5|     c = 3;
  6| te#st`A
  7| ${a}
  8| B
Original: 27 (4, 2)
  2| C`;
  3| let a=1,b=2,c=3;
  4| te#st`A
  5| ${a}
  6| B

Formatted: 44 (6, 3)
  4|     b = 2,
  5|     c = 3;
  6| tes#t`A
  7| ${a}
  8| B
Original: 28 (4, 3)
  2| C`;
  3| let a=1,b=2,c=3;
  4| tes#t`A
  5| ${a}
  6| B

Formatted: 45 (6, 4)
  4|     b = 2,
  5|     c = 3;
  6| test#`A
  7| ${a}
  8| B
Original: 29 (4, 4)
  2| C`;
  3| let a=1,b=2,c=3;
  4| test#`A
  5| ${a}
  6| B

Formatted: 46 (6, 5)
  4|     b = 2,
  5|     c = 3;
  6| test`#A
  7| ${a}
  8| B
Original: 30 (4, 5)
  2| C`;
  3| let a=1,b=2,c=3;
  4| test`#A
  5| ${a}
  6| B

Formatted: 47 (6, 6)
  4|     b = 2,
  5|     c = 3;
  6| test`A#
  7| ${a}
  8| B
Original: 31 (4, 6)
  2| C`;
  3| let a=1,b=2,c=3;
  4| test`A#
  5| ${a}
  6| B

Formatted: 48 (7, 0)
  5|     c = 3;
  6| test`A
  7| #${a}
  8| B
  9| ${b}
Original: 32 (5, 0)
  3| let a=1,b=2,c=3;
  4| test`A
  5| #${a}
  6| B
  7| ${b}

Formatted: 49 (7, 1)
  5|     c = 3;
  6| test`A
  7| $#{a}
  8| B
  9| ${b}
Original: 33 (5, 1)
  3| let a=1,b=2,c=3;
  4| test`A
  5| $#{a}
  6| B
  7| ${b}

Formatted: 50 (7, 2)
  5|     c = 3;
  6| test`A
  7| ${#a}
  8| B
  9| ${b}
Original: 34 (5, 2)
  3| let a=1,b=2,c=3;
  4| test`A
  5| ${#a}
  6| B
  7| ${b}

Formatted: 51 (7, 3)
  5|     c = 3;
  6| test`A
  7| ${a#}
  8| B
  9| ${b}
Original: 35 (5, 3)
  3| let a=1,b=2,c=3;
  4| test`A
  5| ${a#}
  6| B
  7| ${b}

Formatted: 52 (7, 4)
  5|     c = 3;
  6| test`A
  7| ${a}#
  8| B
  9| ${b}
Original: 36 (5, 4)
  3| let a=1,b=2,c=3;
  4| test`A
  5| ${a}#
  6| B
  7| ${b}

Formatted: 53 (8, 0)
  6| test`A
  7| ${a}
  8| #B
  9| ${b}
 10| C
Original: 37 (6, 0)
  4| test`A
  5| ${a}
  6| #B
  7| ${b}
  8| C

Formatted: 54 (8, 1)
  6| test`A
  7| ${a}
  8| B#
  9| ${b}
 10| C
Original: 38 (6, 1)
  4| test`A
  5| ${a}
  6| B#
  7| ${b}
  8| C

Formatted: 55 (9, 0)
  7| ${a}
  8| B
  9| #${b}
 10| C
 11| ${c}`;
Original: 39 (7, 0)
  5| ${a}
  6| B
  7| #${b}
  8| C
  9| ${c}`;

Formatted: 56 (9, 1)
  7| ${a}
  8| B
  9| $#{b}
 10| C
 11| ${c}`;
Original: 40 (7, 1)
  5| ${a}
  6| B
  7| $#{b}
  8| C
  9| ${c}`;

Formatted: 57 (9, 2)
  7| ${a}
  8| B
  9| ${#b}
 10| C
 11| ${c}`;
Original: 41 (7, 2)
  5| ${a}
  6| B
  7| ${#b}
  8| C
  9| ${c}`;

Formatted: 58 (9, 3)
  7| ${a}
  8| B
  9| ${b#}
 10| C
 11| ${c}`;
Original: 42 (7, 3)
  5| ${a}
  6| B
  7| ${b#}
  8| C
  9| ${c}`;

Formatted: 59 (9, 4)
  7| ${a}
  8| B
  9| ${b}#
 10| C
 11| ${c}`;
Original: 43 (7, 4)
  5| ${a}
  6| B
  7| ${b}#
  8| C
  9| ${c}`;

Formatted: 60 (10, 0)
  8| B
  9| ${b}
 10| #C
 11| ${c}`;
 12| `ABC`;
Original: 44 (8, 0)
  6| B
  7| ${b}
  8| #C
  9| ${c}`;
 10| `ABC`;

Formatted: 61 (10, 1)
  8| B
  9| ${b}
 10| C#
 11| ${c}`;
 12| `ABC`;
Original: 45 (8, 1)
  6| B
  7| ${b}
  8| C#
  9| ${c}`;
 10| `ABC`;

Formatted: 62 (11, 0)
  9| ${b}
 10| C
 11| #${c}`;
 12| `ABC`;
 13| a = 1,
Original: 46 (9, 0)
  7| ${b}
  8| C
  9| #${c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 63 (11, 1)
  9| ${b}
 10| C
 11| $#{c}`;
 12| `ABC`;
 13| a = 1,
Original: 47 (9, 1)
  7| ${b}
  8| C
  9| $#{c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 64 (11, 2)
  9| ${b}
 10| C
 11| ${#c}`;
 12| `ABC`;
 13| a = 1,
Original: 48 (9, 2)
  7| ${b}
  8| C
  9| ${#c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 65 (11, 3)
  9| ${b}
 10| C
 11| ${c#}`;
 12| `ABC`;
 13| a = 1,
Original: 49 (9, 3)
  7| ${b}
  8| C
  9| ${c#}`;
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 66 (11, 4)
  9| ${b}
 10| C
 11| ${c}#`;
 12| `ABC`;
 13| a = 1,
Original: 50 (9, 4)
  7| ${b}
  8| C
  9| ${c}#`;
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 67 (11, 5)
  9| ${b}
 10| C
 11| ${c}`#;
 12| `ABC`;
 13| a = 1,
Original: 51 (9, 5)
  7| ${b}
  8| C
  9| ${c}`#;
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 68 (11, 6)
  9| ${b}
 10| C
 11| ${c}`;#
 12| `ABC`;
 13| a = 1,
Original: 52 (9, 6)
  7| ${b}
  8| C
  9| ${c}`;#
 10| `ABC`;
 11| a=1,b=2,c=3;

Formatted: 69 (12, 0)
 10| C
 11| ${c}`;
 12| #`ABC`;
 13| a = 1,
 14| b = 2,
Original: 53 (10, 0)
  8| C
  9| ${c}`;
 10| #`ABC`;
 11| a=1,b=2,c=3;
 12|

Formatted: 70 (12, 1)
 10| C
 11| ${c}`;
 12| `#ABC`;
 13| a = 1,
 14| b = 2,
Original: 54 (10, 1)
  8| C
  9| ${c}`;
 10| `#ABC`;
 11| a=1,b=2,c=3;
 12|

Formatted: 71 (12, 2)
 10| C
 11| ${c}`;
 12| `A#BC`;
 13| a = 1,
 14| b = 2,
Original: 55 (10, 2)
  8| C
  9| ${c}`;
 10| `A#BC`;
 11| a=1,b=2,c=3;
 12|

Formatted: 72 (12, 3)
 10| C
 11| ${c}`;
 12| `AB#C`;
 13| a = 1,
 14| b = 2,
Original: 56 (10, 3)
  8| C
  9| ${c}`;
 10| `AB#C`;
 11| a=1,b=2,c=3;
 12|

Formatted: 73 (12, 4)
 10| C
 11| ${c}`;
 12| `ABC#`;
 13| a = 1,
 14| b = 2,
Original: 57 (10, 4)
  8| C
  9| ${c}`;
 10| `ABC#`;
 11| a=1,b=2,c=3;
 12|

Formatted: 74 (12, 5)
 10| C
 11| ${c}`;
 12| `ABC`#;
 13| a = 1,
 14| b = 2,
Original: 58 (10, 5)
  8| C
  9| ${c}`;
 10| `ABC`#;
 11| a=1,b=2,c=3;
 12|

Formatted: 75 (12, 6)
 10| C
 11| ${c}`;
 12| `ABC`;#
 13| a = 1,
 14| b = 2,
Original: 59 (10, 6)
  8| C
  9| ${c}`;
 10| `ABC`;#
 11| a=1,b=2,c=3;
 12|

Formatted: 76 (13, 0)
 11| ${c}`;
 12| `ABC`;
 13| #a = 1,
 14| b = 2,
 15| c = 3;
Original: 60 (11, 0)
  9| ${c}`;
 10| `ABC`;
 11| #a=1,b=2,c=3;
 12|

Formatted: 77 (13, 1)
 11| ${c}`;
 12| `ABC`;
 13| a# = 1,
 14| b = 2,
 15| c = 3;
Original: 61 (11, 1)
  9| ${c}`;
 10| `ABC`;
 11| a#=1,b=2,c=3;
 12|

Formatted: 78 (13, 2)
 11| ${c}`;
 12| `ABC`;
 13| a #= 1,
 14| b = 2,
 15| c = 3;
Original: 62 (11, 1)
  9| ${c}`;
 10| `ABC`;
 11| a#=1,b=2,c=3;
 12|

Formatted: 79 (13, 3)
 11| ${c}`;
 12| `ABC`;
 13| a =# 1,
 14| b = 2,
 15| c = 3;
Original: 62 (11, 2)
  9| ${c}`;
 10| `ABC`;
 11| a=#1,b=2,c=3;
 12|

Formatted: 80 (13, 4)
 11| ${c}`;
 12| `ABC`;
 13| a = #1,
 14| b = 2,
 15| c = 3;
Original: 63 (11, 2)
  9| ${c}`;
 10| `ABC`;
 11| a=#1,b=2,c=3;
 12|

Formatted: 81 (13, 5)
 11| ${c}`;
 12| `ABC`;
 13| a = 1#,
 14| b = 2,
 15| c = 3;
Original: 63 (11, 3)
  9| ${c}`;
 10| `ABC`;
 11| a=1#,b=2,c=3;
 12|

Formatted: 82 (13, 6)
 11| ${c}`;
 12| `ABC`;
 13| a = 1,#
 14| b = 2,
 15| c = 3;
Original: 64 (11, 4)
  9| ${c}`;
 10| `ABC`;
 11| a=1,#b=2,c=3;
 12|

Formatted: 83 (14, 0)
 12| `ABC`;
 13| a = 1,
 14| #b = 2,
 15| c = 3;
 16|
Original: 65 (11, 4)
  9| ${c}`;
 10| `ABC`;
 11| a=1,#b=2,c=3;
 12|

Formatted: 84 (14, 1)
 12| `ABC`;
 13| a = 1,
 14| b# = 2,
 15| c = 3;
 16|
Original: 65 (11, 5)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b#=2,c=3;
 12|

Formatted: 85 (14, 2)
 12| `ABC`;
 13| a = 1,
 14| b #= 2,
 15| c = 3;
 16|
Original: 66 (11, 5)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b#=2,c=3;
 12|

Formatted: 86 (14, 3)
 12| `ABC`;
 13| a = 1,
 14| b =# 2,
 15| c = 3;
 16|
Original: 66 (11, 6)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=#2,c=3;
 12|

Formatted: 87 (14, 4)
 12| `ABC`;
 13| a = 1,
 14| b = #2,
 15| c = 3;
 16|
Original: 67 (11, 6)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=#2,c=3;
 12|

Formatted: 88 (14, 5)
 12| `ABC`;
 13| a = 1,
 14| b = 2#,
 15| c = 3;
 16|
Original: 67 (11, 7)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2#,c=3;
 12|

Formatted: 89 (14, 6)
 12| `ABC`;
 13| a = 1,
 14| b = 2,#
 15| c = 3;
 16|
Original: 68 (11, 8)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,#c=3;
 12|

Formatted: 90 (15, 0)
 13| a = 1,
 14| b = 2,
 15| #c = 3;
 16|
Original: 69 (11, 8)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,#c=3;
 12|

Formatted: 91 (15, 1)
 13| a = 1,
 14| b = 2,
 15| c# = 3;
 16|
Original: 69 (11, 9)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c#=3;
 12|

Formatted: 92 (15, 2)
 13| a = 1,
 14| b = 2,
 15| c #= 3;
 16|
Original: 70 (11, 9)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c#=3;
 12|

Formatted: 93 (15, 3)
 13| a = 1,
 14| b = 2,
 15| c =# 3;
 16|
Original: 70 (11, 10)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=#3;
 12|

Formatted: 94 (15, 4)
 13| a = 1,
 14| b = 2,
 15| c = #3;
 16|
Original: 71 (11, 10)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=#3;
 12|

Formatted: 95 (15, 5)
 13| a = 1,
 14| b = 2,
 15| c = 3#;
 16|
Original: 71 (11, 11)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=3#;
 12|

Formatted: 96 (15, 6)
 13| a = 1,
 14| b = 2,
 15| c = 3;#
 16|
Original: 72 (11, 12)
  9| ${c}`;
 10| `ABC`;
 11| a=1,b=2,c=3;#
 12|

Formatted: 97 (16, 0)
 14| b = 2,
 15| c = 3;
 16| #
Original: 73 (12, 0)
 10| `ABC`;
 11| a=1,b=2,c=3;
 12| #
